requirements.txt
opencv-python>=4.5.0
matplotlib>=3.5.0
numpy>=1.21.0
PyYAML>=6.0
ultralytics>=8.0.0
